+++ src/reactpy_django/utils.py
@@ -12,7 +12,6 @@ from inspect import iscoroutinefunction
 from typing import Any, Callable, Sequence
 
 from channels.db import database_sync_to_async
-from django.core.cache import caches
 from django.db.models import ManyToManyField, ManyToOneRel, prefetch_related_objects
 from django.db.models.base import Model
 from django.db.models.query import QuerySet
@@ -23,7 +22,6 @@ from django.utils.encoding import smart_str
 from django.views import View
 
 from reactpy_django.exceptions import ComponentDoesNotExistError, ComponentParamError
-
 
 _logger = logging.getLogger(__name__)
 _component_tag = r"(?P<tag>component)"
@@ -332,39 +330,22 @@ def create_cache_key(*args):
 def db_cleanup(immediate: bool = False):
     """Deletes expired component sessions from the database.
     This function may be expanded in the future to include additional cleanup tasks."""
-    from .config import (
-        REACTPY_CACHE,
-        REACTPY_DATABASE,
-        REACTPY_DEBUG_MODE,
-        REACTPY_RECONNECT_MAX,
-    )
-    from .models import ComponentSession
+    from .config import REACTPY_DATABASE, REACTPY_DEBUG_MODE, REACTPY_RECONNECT_MAX
+    from .models import ComponentSession, Config
 
     clean_started_at = datetime.now()
-    cache_key: str = create_cache_key("last_cleaned")
-    now_str: str = datetime.strftime(timezone.now(), DATE_FORMAT)
-    cleaned_at_str: str = caches[REACTPY_CACHE].get(cache_key)
-    cleaned_at: datetime = timezone.make_aware(
-        datetime.strptime(cleaned_at_str or now_str, DATE_FORMAT)
-    )
+    config = Config.load()
+    cleaned_at = config.cleaned_at
     clean_needed_by = cleaned_at + timedelta(seconds=REACTPY_RECONNECT_MAX)
-    expires_by: datetime = timezone.now() - timedelta(seconds=REACTPY_RECONNECT_MAX)
-
-    # Component params exist in the DB, but we don't know when they were last cleaned
-    if not cleaned_at_str and ComponentSession.objects.using(REACTPY_DATABASE).all():
-        _logger.warning(
-            "ReactPy has detected component sessions in the database, "
-            "but no timestamp was found in cache. This may indicate that "
-            "the cache has been cleared."
-        )
+    expires_by = timezone.now() - timedelta(seconds=REACTPY_RECONNECT_MAX)
 
     # Delete expired component parameters
-    # Use timestamps in cache (`cleaned_at_str`) as a no-dependency rate limiter
-    if immediate or not cleaned_at_str or timezone.now() >= clean_needed_by:
+    if immediate or timezone.now() >= clean_needed_by:
         ComponentSession.objects.using(REACTPY_DATABASE).filter(
             last_accessed__lte=expires_by
         ).delete()
-        caches[REACTPY_CACHE].set(cache_key, now_str, timeout=None)
+        config.cleaned_at = timezone.now()
+        config.save()
 
     # Check if cleaning took abnormally long
     clean_duration = datetime.now() - clean_started_at